get and post product

+++ database/migrations/2020_04_26_082922_create_orders_table.php
@@ -14,8 +14,8 @@ class CreateOrdersTable extends Migration
     public function up()
     {
         Schema::create('orders', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('user_id')->unsigned();
+            $table->increments('id');
+            $table->integer('user_id')->unsigned();
             $table->string('name');
             $table->text('description');
             $table->integer('price')->nullable();
